--- src/test/java/com/historycode/cucumber/steps/NewsPageSteps.java
@@ -1,10 +1,0 @@
-package com.historycode.cucumber.steps;
-
-import com.historycode.ui.page.adminpanel.newspage.modal.CreateEditNewsModal;
-import io.cucumber.java.en.And;
-
-public class NewsPageSteps extends BaseAdminPanelSteps {
-    protected CreateEditNewsModal createEditNewsModal;
-
-
-}
