--- src/test/java/com/historycode/cucumber/steps/PartnerPageSteps.java
@@ -1,7 +1,0 @@
-package com.historycode.cucumber.steps;
-
-import com.historycode.ui.page.adminpanel.partnerspage.modal.CreatePartnersModal;
-
-public class PartnerPageSteps extends BaseAdminPanelSteps{
-    protected CreatePartnersModal createPartnersModal;
-}
